+++ ChangeLog.md
@@ -8,6 +8,8 @@
 
     - Write 'moveDir' schema (todo #110)
     - Define test/sh for ttp.pl vars (todo #126)
+    - Define test/sh for ttp.pl vars (todo #126)
+    - Define test/sh for daemon.pl vars (todo #127)
 
 ### 4.13.0